% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesCR.R
\docType{package}
\name{BayesCR}
\alias{BayesCR}
\alias{BayesCR-package}
\alias{BayesCR-package}
\title{BayesCR: Bayesian Analysis of Censored Regression Models Under Scale Mixture of Skew Normal Distributions}
\description{
Propose a parametric fit for censored linear regression models based on SMSN distributions, from a Bayesian perspective. Also, generates SMSN random variables.
}
\references{
Monique B. Massuia, Aldo M. Garay, Celso R. Cabral and Victor H. Lachos. "Bayesian analysis of censored linear regression models with scale mixtures of skew-normal distributions". Statistics and Its Interface, 2017, vol. 10, pages 425-439
}
\seealso{
\code{\link{Bayes.CR}}, \code{\link{rSMSN}}, \code{\link{motorettes}}
}
\author{
Aldo M. Garay \email{medina_garay@yahoo.com}, Monique B. Massuia \email{moniquemassuia@gmail.com}, 
Victor H. Lachos \email{hlachos@ime.unicamp.br} and Eraldo B. Anjos Filho \email{ebdaf1@de.ufpe.br}

Maintainer: Aldo M. Garay \email{medina_garay@yahoo.com}
}
